Initial commit; created separate status_monitor module; moved status_util to status.py; adjusted unit tests; added codemeta.json
